Use long long instead of int64_t
int64_t appears to have portability issues with regard to printf formatting
specifiers. On some systems it's long long; on some others it's just long. This
can be worked around by using PRI64 definition, but it looks ugly; it's much
simpler to just explicitly use long long type for printing

+++ cpu6.c
@@ -557,7 +557,7 @@ static int bignum_op() {
 
 
 		// Convert to little endian
-		uint64_t num = 0;
+		unsigned long long num = 0;
 		for (int i=0; i < b_size; i++) {
 			num = num << 8 | mmu_mem_read8(src_addr+i);
 		}
